move feature components
